# apps/api/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/votes": {
      "get": {
        "operationId": "VotesController_getVotes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns a list of votes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VoteListItemResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      },
      "post": {
        "operationId": "VotesController_createVote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns the created vote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVoteResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}": {
      "get": {
        "operationId": "VotesController_getVoteDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the vote detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      },
      "patch": {
        "operationId": "VotesController_updateVote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the updated vote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVoteResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}/voter-status": {
      "get": {
        "operationId": "VotesController_getVoterStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the voter status for the given vote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoterStatusResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}/ruleset": {
      "put": {
        "operationId": "VotesController_setVoteRuleset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVoteRulesetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the configured vote ruleset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetVoteRulesetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}/questions": {
      "post": {
        "operationId": "VotesController_createVoteQuestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoteQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question created"
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}/questions/{questionId}": {
      "patch": {
        "operationId": "VotesController_updateVoteQuestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVoteQuestionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Question updated"
          }
        },
        "tags": [
          "Votes"
        ]
      },
      "delete": {
        "operationId": "VotesController_deleteVoteQuestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Question deleted"
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}/schedule": {
      "post": {
        "operationId": "VotesController_scheduleVote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the scheduled vote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVoteResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}/delegation-candidates": {
      "get": {
        "operationId": "VotesController_getDelegationCandidates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unitId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a list of memberships eligible for delegation for this unit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DelegationCandidateDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Votes"
        ]
      }
    },
    "/api/votes/{id}/consents": {
      "post": {
        "operationId": "VotesController_createVoteConsent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoteConsentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Records a delegation consent for a specific unit"
          }
        },
        "tags": [
          "Votes"
        ]
      }
    }
  },
  "info": {
    "title": "HOA Manager API",
    "description": "The HOA Manager API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "VoterSummaryDto": {
        "type": "object",
        "properties": {
          "canVote": {
            "type": "boolean"
          },
          "requiresDelegation": {
            "type": "boolean"
          },
          "isDelegated": {
            "type": "boolean"
          }
        },
        "required": [
          "canVote",
          "requiresDelegation",
          "isDelegated"
        ]
      },
      "VoteListItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "scheduledFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "scheduledTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "voterSummary": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/VoterSummaryDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "title",
          "status"
        ]
      },
      "SetVoteRulesetResponseDto": {
        "type": "object",
        "properties": {
          "weightBasis": {
            "type": "string",
            "enum": [
              "UNIT_SHARE",
              "ONE_UNIT_ONE_VOTE"
            ]
          },
          "quorumMeasure": {
            "type": "string",
            "enum": [
              "UNIT_SHARE",
              "UNIT_COUNT"
            ]
          },
          "quorumElectorateBasis": {
            "type": "string",
            "enum": [
              "ALL_UNITS",
              "ELIGIBLE_UNITS_ONLY"
            ]
          },
          "quorumThreshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "majorityRuleType": {
            "type": "string",
            "enum": [
              "SIMPLE_MAJORITY",
              "QUALIFIED_MAJORITY"
            ]
          },
          "majorityThreshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "allowAbstain": {
            "type": "boolean"
          },
          "abstainExcludedFromMajorityDenominator": {
            "type": "boolean"
          },
          "allowCoOwnerIndividualVote": {
            "type": "boolean"
          }
        },
        "required": [
          "weightBasis",
          "quorumMeasure",
          "quorumElectorateBasis",
          "quorumThreshold",
          "majorityRuleType",
          "allowAbstain",
          "abstainExcludedFromMajorityDenominator",
          "allowCoOwnerIndividualVote"
        ]
      },
...
      "VoteDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "scheduledFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "scheduledTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "ruleset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/SetVoteRulesetResponseDto"
              }
            ]
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VoteQuestionResponseDto"
            }
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "questions"
        ]
      },
      "TotalVotingPowerDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number"
          },
          "maximum": {
            "type": "number"
          }
        },
        "required": [
          "value",
          "maximum"
        ]
      },
...
      "VoterStatusResponseDto": {
        "type": "object",
        "properties": {
          "canVote": {
            "type": "boolean"
          },
          "totalVotingPower": {
            "$ref": "#/components/schemas/TotalVotingPowerDto"
          },
          "owningUnits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwningUnitStatusDto"
            }
          }
        },
        "required": [
          "canVote",
          "totalVotingPower",
          "owningUnits"
        ]
      },
      "CreateVoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "scheduledFrom": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "scheduledTo": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "title"
        ]
      },
      "CreateVoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "scheduledFrom": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "scheduledTo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status"
        ]
      },
      "SetVoteRulesetDto": {
        "type": "object",
        "properties": {
          "weightBasis": {
            "type": "string",
            "enum": [
              "UNIT_SHARE",
              "ONE_UNIT_ONE_VOTE"
            ]
          },
          "quorumMeasure": {
            "type": "string",
            "enum": [
              "UNIT_SHARE",
              "UNIT_COUNT"
            ]
          },
          "quorumElectorateBasis": {
            "type": "string",
            "enum": [
              "ALL_UNITS",
              "ELIGIBLE_UNITS_ONLY"
            ]
          },
          "quorumThreshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "majorityRuleType": {
            "type": "string",
            "enum": [
              "SIMPLE_MAJORITY",
              "QUALIFIED_MAJORITY"
            ]
          },
          "majorityThreshold": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "allowAbstain": {
            "type": "boolean"
          },
          "abstainExcludedFromMajorityDenominator": {
            "type": "boolean"
          },
          "allowCoOwnerIndividualVote": {
            "type": "boolean"
          }
        },
        "required": [
          "weightBasis",
          "quorumMeasure",
          "quorumElectorateBasis",
          "quorumThreshold",
          "majorityRuleType",
          "allowAbstain",
          "abstainExcludedFromMajorityDenominator",
          "allowCoOwnerIndividualVote"
        ]
      },
      "CreateVoteQuestionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "YES_NO",
              "SINGLE_CHOICE"
            ]
          },
          "sortOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string",
                  "minLength": 1
                },
                "sortOrder": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "label"
              ]
            }
          },
          "rulesetOverride": {
            "type": "object",
            "properties": {
              "weightBasis": {
                "type": "string",
                "enum": [
                  "UNIT_SHARE",
                  "ONE_UNIT_ONE_VOTE"
                ]
              },
              "quorumMeasure": {
                "type": "string",
                "enum": [
                  "UNIT_SHARE",
                  "UNIT_COUNT"
                ]
              },
              "quorumElectorateBasis": {
                "type": "string",
                "enum": [
                  "ALL_UNITS",
                  "ELIGIBLE_UNITS_ONLY"
                ]
              },
              "quorumThreshold": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "majorityRuleType": {
                "type": "string",
                "enum": [
                  "SIMPLE_MAJORITY",
                  "QUALIFIED_MAJORITY"
                ]
              },
              "majorityThreshold": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "allowAbstain": {
                "type": "boolean"
              },
              "abstainExcludedFromMajorityDenominator": {
                "type": "boolean"
              },
              "allowCoOwnerIndividualVote": {
                "type": "boolean"
              }
            },
            "required": [
              "weightBasis",
              "quorumMeasure",
              "quorumElectorateBasis",
              "quorumThreshold",
              "majorityRuleType",
              "allowAbstain",
              "abstainExcludedFromMajorityDenominator",
              "allowCoOwnerIndividualVote"
            ],
            "x-nestjs_zod-self-required": false
          }
        },
        "required": [
          "title",
          "type"
        ]
      },
      "UpdateVoteQuestionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "YES_NO",
              "SINGLE_CHOICE"
            ]
          },
          "sortOrder": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "options": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string",
                  "minLength": 1
                },
                "sortOrder": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "label"
              ]
            }
          },
          "rulesetOverride": {
            "type": "object",
            "properties": {
              "weightBasis": {
                "type": "string",
                "enum": [
                  "UNIT_SHARE",
                  "ONE_UNIT_ONE_VOTE"
                ]
              },
              "quorumMeasure": {
                "type": "string",
                "enum": [
                  "UNIT_SHARE",
                  "UNIT_COUNT"
                ]
              },
              "quorumElectorateBasis": {
                "type": "string",
                "enum": [
                  "ALL_UNITS",
                  "ELIGIBLE_UNITS_ONLY"
                ]
              },
              "quorumThreshold": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "majorityRuleType": {
                "type": "string",
                "enum": [
                  "SIMPLE_MAJORITY",
                  "QUALIFIED_MAJORITY"
                ]
              },
              "majorityThreshold": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "allowAbstain": {
                "type": "boolean"
              },
              "abstainExcludedFromMajorityDenominator": {
                "type": "boolean"
              },
              "allowCoOwnerIndividualVote": {
                "type": "boolean"
              }
            },
            "required": [
              "weightBasis",
              "quorumMeasure",
              "quorumElectorateBasis",
              "quorumThreshold",
              "majorityRuleType",
              "allowAbstain",
              "abstainExcludedFromMajorityDenominator",
              "allowCoOwnerIndividualVote"
            ],
            "x-nestjs_zod-self-required": false
          }
        },
        "required": [
          "title",
          "type"
        ]
      },
      "UpdateVoteDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "scheduledFrom": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "scheduledTo": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "title"
        ]
      },
      "DelegationCandidateDto": {
        "type": "object",
        "properties": {
          "membershipId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "hasDelegatedToRequester": {
            "type": "boolean"
          },
          "isEligible": {
            "type": "boolean"
          }
        },
        "required": [
          "membershipId",
          "name",
          "hasDelegatedToRequester",
          "isEligible"
        ]
      },
      "CreateVoteConsentDto": {
        "type": "object",
        "properties": {
          "unitId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "delegateMembershipId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "unitId",
          "delegateMembershipId"
        ]
      }
    }
  }
}