# apps/portal/src/api/generated/model/delegationCandidateDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */

export interface DelegationCandidateDto {
    membershipId: string;
    name: string;
    hasDelegatedToRequester: boolean;
    isEligible: boolean;
}